Don't reassign int value from `int64`
This was caught by CodeQL scanning and saves reassigning the `int64` value down to just an `int`
`int64` is what Miniflux stores its entry IDs as so matches what we're given

+++ store/sqlite/sqlite.go
@@ -105,7 +105,7 @@ func (d db) GetEntries() ([]models.Message, error) {
 	return results, nil
 }
 
-func (d db) DeleteEntryByID(id int) error {
+func (d db) DeleteEntryByID(id int64) error {
 	_, err := d.ctx.Exec(`
 	DELETE from entries where id=?
 	`, id)
